common: move ThreadingDialogs and wxAppWithHelpers to main project
