+++ pcsx2/gui/wxAppWithHelpers.h
